Refactor: UserGetRes file structure
